+++ formula-android/src/main/java/com/instacart/formula/android/FeatureFactory.kt
@@ -26,8 +26,8 @@ package com.instacart.formula.android
  * }
  * ```
  *
- * Once we define a [FeatureFactory], we need to [bind][FragmentStoreBuilder.bind] it to a
- * [FragmentFlowStore]. The fragment flow store will call [initialize] the first time
+ * Once we define a [FeatureFactory], we need to [bind][FeaturesBuilder.bind] it to a
+ * [FragmentStore]. The fragment flow store will call [initialize] the first time
  * [FormulaFragment] with a new [Key] is attached. It will subscribe to the state management
  * and persist it across configuration changes.
  *