+++ formula-android/src/main/java/com/instacart/formula/android/internal/ActivityManager.kt
@@ -43,11 +43,11 @@ internal class ActivityManager<Activity : FragmentActivity>(
             activity = activity,
             fragmentEnvironment = environment,
             onLifecycleEvent = {
-                store.contracts.onLifecycleEffect(it)
+                store.fragmentStore.onLifecycleEffect(it)
                 store.onFragmentLifecycleEvent?.invoke(it)
             },
             onLifecycleState = delegate::updateFragmentLifecycleState,
-            onFragmentViewStateChanged = store.contracts::onVisibilityChanged
+            onFragmentViewStateChanged = store.fragmentStore::onVisibilityChanged
         )
     }
 
@@ -56,7 +56,7 @@ internal class ActivityManager<Activity : FragmentActivity>(
         delegate.onLifecycleStateChanged(Lifecycle.State.CREATED)
         val renderView = fragmentRenderView ?: throw callOnPreCreateException(activity)
 
-        uiSubscription = delegate.fragmentFlowState().subscribe {
+        uiSubscription = delegate.fragmentState().subscribe {
             renderView.render(it)
             store.onRenderFragmentState?.invoke(activity, it)
         }
@@ -108,9 +108,9 @@ internal class ActivityManager<Activity : FragmentActivity>(
 
     private fun subscribeToFragmentStateChanges(): Disposable {
         return store
-            .contracts
+            .fragmentStore
             .state(environment)
-            .subscribe(delegate.fragmentFlowStateRelay::accept)
+            .subscribe(delegate.fragmentStateRelay::accept)
     }
 
     private fun callOnPreCreateException(activity: FragmentActivity): IllegalStateException {