+++ formula-android/src/test/java/com/instacart/formula/android/fakes/FakeComponent.kt
@@ -1,25 +1,14 @@
 package com.instacart.formula.android.fakes
 
 import com.instacart.formula.android.FragmentKey
-import com.instacart.formula.android.DisposableScope
 import com.jakewharton.rxrelay3.PublishRelay
 import io.reactivex.rxjava3.core.Observable
 
 class FakeComponent {
-    val initialized = mutableListOf<Pair<FakeAuthFlowFactory.Component, FragmentKey>>()
     val updateRelay: PublishRelay<Pair<FragmentKey, String>> = PublishRelay.create()
 
     fun state(key: FragmentKey): Observable<String> {
         val updates = updateRelay.filter { it.first == key }.map { it.second }
         return updates.startWithItem("${key.tag}-state")
-    }
-
-    fun createAuthFlowComponent(): DisposableScope<FakeAuthFlowFactory.Component> {
-        val component = FakeAuthFlowFactory.Component(onInitialized = { component, key ->
-            initialized.add(component to key)
-        })
-        return DisposableScope(component, {
-            initialized.clear()
-        })
     }
 }